requirements/local.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/local.in
#
alabaster==0.7.13
    # via sphinx
amqp==5.1.1
    # via kombu
anyio==3.6.2
    # via watchfiles
argon2-cffi==21.3.0
    # via -r requirements\base.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.6.0
    # via django
astroid==2.15.1
    # via
    #   pylint
    #   pylint-celery
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via
    #   jsonschema
    #   pytest
babel==2.11.0
    # via sphinx
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
black==23.1.0
    # via -r requirements/local.in
celery==5.2.7
    # via
    #   -r requirements\base.in
    #   django-celery-beat
    #   flower
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   ipython
    #   pylint
    #   pytest
    #   sphinx
    #   sphinx-autobuild
coverage[toml]==7.2.2
    # via
    #   -r requirements/local.in
    #   django-coverage-plugin
    #   pytest-cov
cron-descriptor==1.2.35
    # via django-celery-beat
cryptography==39.0.0
    # via pyjwt
debugpy==1.6.6
    # via -r requirements/local.in
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via python3-openid
dill==0.3.6
    # via pylint
distlib==0.3.6
    # via virtualenv
dj-rest-auth[with_social]==2.2.8
    # via -r requirements\base.in
django==4.1.7
    # via
    #   -r requirements\base.in
    #   dj-rest-auth
    #   django-allauth
    #   django-celery-beat
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   sphinxcontrib-django
django-allauth==0.52.0
    # via dj-rest-auth
django-celery-beat==2.5.0
    # via -r requirements\base.in
django-cors-headers==3.14.0
    # via -r requirements\base.in
django-coverage-plugin==3.0.0
    # via -r requirements/local.in
django-debug-toolbar==3.8.1
    # via -r requirements/local.in
django-environ==0.10.0
    # via -r requirements\base.in
django-extensions==3.2.1
    # via -r requirements\base.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via
    #   -r requirements\base.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.2.2
    # via -r requirements\base.in
docutils==0.19
    # via sphinx
drf-spectacular==0.26.1
    # via -r requirements\base.in
execnet==1.9.0
    # via pytest-xdist
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via -r requirements/local.in
faker==16.6.1
    # via factory-boy
filelock==3.9.0
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements/local.in
    #   flake8-isort
flake8-isort==6.0.0
    # via -r requirements/local.in
flower==1.2.0
    # via -r requirements\base.in
humanize==4.4.0
    # via flower
identify==2.5.15
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   requests
imagesize==1.4.1
    # via sphinx
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
ipdb==0.13.11
    # via -r requirements/local.in
ipython==8.8.0
    # via ipdb
isort==5.11.4
    # via
    #   flake8-isort
    #   pylint
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via sphinx
jsonschema==4.17.3
    # via drf-spectacular
kombu==5.2.4
    # via celery
lazy-object-proxy==1.9.0
    # via astroid
livereload==2.6.3
    # via sphinx-autobuild
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy-extensions==0.4.3
    # via black
nodeenv==1.7.0
    # via pre-commit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.0
    # via
    #   black
    #   pytest
    #   pytest-sugar
    #   sphinx
parso==0.8.3
    # via jedi
pathspec==0.10.3
    # via black
pickleshare==0.7.5
    # via ipython
platformdirs==2.6.2
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via pytest
pprintpp==0.4.0
    # via sphinxcontrib-django
pre-commit==3.2.1
    # via -r requirements/local.in
prometheus-client==0.16.0
    # via flower
prompt-toolkit==3.0.36
    # via
    #   click-repl
    #   ipython
psycopg2==2.9.5
    # via -r requirements\base.in
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pygments==2.14.0
    # via
    #   ipython
    #   sphinx
pyjwt[crypto]==2.6.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
pylint==2.17.1
    # via
    #   -r requirements/local.in
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via -r requirements/local.in
pylint-django==2.5.3
    # via -r requirements/local.in
pylint-plugin-utils==0.7
    # via
    #   pylint-celery
    #   pylint-django
pyrsistent==0.19.3
    # via jsonschema
pytest==7.2.2
    # via
    #   -r requirements/local.in
    #   pytest-cov
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-cov==4.0.0
    # via -r requirements/local.in
pytest-django==4.5.2
    # via -r requirements/local.in
pytest-sugar==0.9.6
    # via -r requirements/local.in
pytest-xdist==3.2.1
    # via -r requirements/local.in
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   faker
    #   python-crontab
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7.1
    # via
    #   babel
    #   celery
    #   django-timezone-field
    #   djangorestframework
    #   flower
pyyaml==6.0
    # via
    #   drf-spectacular
    #   pre-commit
requests==2.28.2
    # via
    #   django-allauth
    #   requests-oauthlib
    #   sphinx
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via
    #   asttokens
    #   click-repl
    #   livereload
    #   python-dateutil
sniffio==1.3.0
    # via anyio
snowballstemmer==2.2.0
    # via sphinx
sphinx==6.1.3
    # via
    #   -r requirements/local.in
    #   sphinx-autobuild
    #   sphinxcontrib-django
sphinx-autobuild==2021.3.14
    # via -r requirements/local.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-django==2.2
    # via -r requirements/local.in
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
termcolor==2.2.0
    # via pytest-sugar
tomlkit==0.11.6
    # via pylint
tornado==6.2
    # via
    #   flower
    #   livereload
traitlets==5.8.1
    # via
    #   ipython
    #   matplotlib-inline
tzdata==2022.7
    # via
    #   django
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.14
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.17.1
    # via pre-commit
watchdog==2.2.1
    # via werkzeug
watchfiles==0.18.1
    # via -r requirements/local.in
wcwidth==0.2.6
    # via prompt-toolkit
werkzeug[watchdog]==2.2.2
    # via -r requirements/local.in
wrapt==1.14.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
